template: styles: Change colors for dark mode
Change background and font colors for dark mode, both for the slide
itself and for the bottom-panel.
The idea behind changing black into very dark grey and white into very
light grey is to make the dark mode easier on the eyes, so it is
actually a valid option when presenting.
The bottom-panel text color is made always black because white on yellow is
hard to read.

Signed-off-by: Patrick Barsanti <patrick.barsanti@amarulasolutions.com>