template: layouts: Fix two-cols layout
Two-cols layout only uses `slidev-layout` for the left part of the slide.
This means that the overall style, including the background color, is
not rendering correctly in the right part of the slide.

At the same time, the right part of the slide also does not render the
markdown code that should be rendering.

Fix the first issue by moving `slidev-layout` to the container class.
Fix the second issue by changing the template code to match the original
one by slidev for the two-cols layout.
The right side of the slide's slot has the correct name so it is also possible
to use `::right::` syntax sugar instead of specifying the slots when
writing markdown.

Signed-off-by: Patrick Barsanti <patrick.barsanti@amarulasolutions.com>