clk: Add `A clock signal` slide
Add slide titled `A clock signal`, and the images it is using.
This shows the difference between a clock signal seen by a logic
analyzer and seen by an oscilloscope, evidencing all imperfections that
can be normally seen in a clock signal, like overshooting and
non-instantaneous rise and fall times.

Signed-off-by: Patrick Barsanti <patrick.barsanti@amarulasolutions.com>