clk: Add `The clock tree` slide
Add slide titled `The clock tree`, and the image it uses.
This explains the concept of a clock tree, and how it is generated.
The image may not be the best, but it does the job in showing that
starting from a single crystal many different clock frequencies are
generated and distributed to the different parts of the system.

Signed-off-by: Patrick Barsanti <patrick.barsanti@amarulasolutions.com>